requirements.txt
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
loguru==0.5.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pydantic==1.8.2
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
uvicorn==0.15.0
win32-setctime==1.0.3
thefuzz
python-Levenshtein
fastapi
starlette
gunicorn
uvicorn
numpy
python-multipart
omegaconf
einops
gradio
scipy
ocrd-fork-pylsd
